+++ prism.py
@@ -153,7 +153,7 @@ Examples:
         action="store_true",
         help="Disable plugin loading",
     )
-    parser.add_argument("--version", action="version", version="PRISM 1.6.2")
+    parser.add_argument("--version", action="version", version="PRISM 1.6.5")
 
     args = parser.parse_args()